# docs/infrastructure/api_v2/current-openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Assixx API Documentation",
    "version": "1.0.0",
    "description": "Multi-Tenant SaaS Platform für Industrieunternehmen",
    "contact": { "name": "SCS-Technik", "email": "support@scs-technik.de" },
    "license": { "name": "Proprietary", "url": "https://assixx.com/license" }
  },
  "servers": [
    { "url": "https://api.assixx.com/api", "description": "Production Server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error message" },
          "message": {
            "type": "string",
            "description": "Error message (alternative field)"
          },
          "code": { "type": "string", "description": "Error code" },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "description": "User ID" },
          "username": { "type": "string", "description": "Username" },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address"
          },
          "first_name": { "type": "string", "description": "First name" },
          "last_name": { "type": "string", "description": "Last name" },
          "role": {
            "type": "string",
            "enum": ["root", "admin", "employee"],
            "description": "User role"
          },
          "tenant_id": { "type": "integer", "description": "Tenant ID" },
          "department_id": {
            "type": "integer",
            "nullable": true,
            "description": "Department ID"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the user is active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": ["id", "username", "email", "role", "tenant_id"]
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "description": "Authenticate user with username, password and optional browser fingerprint",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username for authentication",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User password",
                    "example": "SecurePass123!"
                  },
                  "fingerprint": {
                    "type": "string",
                    "description": "Browser fingerprint for session isolation",
                    "example": "abc123def456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "description": "HTTP-only cookie containing JWT token"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Login erfolgreich"
                    },
                    "token": {
                      "type": "string",
                      "description": "JWT token (also set as httpOnly cookie)"
                    },
                    "role": {
                      "type": "string",
                      "enum": ["root", "admin", "employee"]
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "username": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string" },
                        "tenant_id": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register new user",
        "description": "Create a new user account with optional tenant assignment",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Unique username",
                    "example": "johndoe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address",
                    "example": "john.doe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Strong password",
                    "example": "SecurePass123!"
                  },
                  "first_name": {
                    "type": "string",
                    "description": "First name",
                    "example": "John"
                  },
                  "last_name": {
                    "type": "string",
                    "description": "Last name",
                    "example": "Doe"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["admin", "employee"],
                    "description": "User role (cannot register as root)",
                    "example": "employee"
                  },
                  "tenant_id": {
                    "type": "integer",
                    "description": "Tenant ID for multi-tenant assignment",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Registration successful"
                    },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout user",
        "description": "Logout the current user and clear authentication token",
        "tags": ["Authentication"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Logout successful",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "description": "Clears the httpOnly token cookie"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logout successful"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}